Change cast type from size_t to uint16_t to prevent data loss

+++ source/ice_api_private.c
@@ -239,7 +239,7 @@ IceResult_t Ice_FinalizeStunPacket( IceContext_t * pContext,
     IceResult_t iceResult = ICE_RESULT_OK;
     StunResult_t stunResult = STUN_RESULT_OK;
     uint8_t messageIntegrity[ STUN_HMAC_VALUE_LENGTH ];
-    size_t messageIntegrityLength = STUN_HMAC_VALUE_LENGTH;
+    uint16_t messageIntegrityLength = STUN_HMAC_VALUE_LENGTH;
     uint8_t * pIntegrityCalculationData = NULL;
     uint16_t integrityCalculationDataLength = 0;
     uint8_t * pFingerprintCalculationData = NULL;
@@ -266,7 +266,7 @@ IceResult_t Ice_FinalizeStunPacket( IceContext_t * pContext,
             {
                 stunResult = StunSerializer_AddAttributeIntegrity( pStunCtx,
                                                                    &( messageIntegrity[ 0 ] ),
-                                                                   ( uint16_t ) messageIntegrityLength );
+                                                                   messageIntegrityLength );
             }
         }
     }
@@ -328,7 +328,7 @@ IceHandleStunPacketResult_t Ice_DeserializeStunPacket( IceContext_t * pContext,
     uint8_t * pIntegrityCalculationData = NULL;
     uint16_t integrityCalculationDataLength = 0;
     uint8_t messageIntegrity[ STUN_HMAC_VALUE_LENGTH ];
-    size_t messageIntegrityLength = STUN_HMAC_VALUE_LENGTH;
+    uint16_t messageIntegrityLength = STUN_HMAC_VALUE_LENGTH;
     uint8_t * pFingerprintCalculationData = NULL;
     uint16_t fingerprintCalculationDataLength = 0;
     uint32_t fingerprint;
@@ -396,7 +396,7 @@ IceHandleStunPacketResult_t Ice_DeserializeStunPacket( IceContext_t * pContext,
                                                                            &( messageIntegrityLength ) );
 
                             if( ( iceResult != ICE_RESULT_OK ) ||
-                                ( messageIntegrityLength != ( size_t ) stunAttribute.attributeValueLength ) ||
+                                ( messageIntegrityLength != stunAttribute.attributeValueLength ) ||
                                 ( memcmp( &( messageIntegrity[ 0 ] ),
                                           stunAttribute.pAttributeValue,
                                           stunAttribute.attributeValueLength ) != 0 ) )
